Correctly outputs the right amount of words to file
There seems to be memory/C++ issues that lead to some nodes that should be registered (and when checked in debug mode, have indeed been registered) having registered = false and not being in the registered vector. This makes no sense as the only nodes left after the dawg is created are registered nodes, the rest are deleted in the process. The only option I seem to be able to find that prints all the possible words while also not printing new made up words is to change nothing except not check whether nodes are registered any more when printing.