Creating 'Compact Trie'
This is a trie made out of two arrays, one for characters/branches and one for 'node pointers' / offsets. Can turn the regular DAWG into a compact DAWG. Cans till search for words and traverse the trie.

It is not yet complete, need to implement more general functionality + frequency aspect + writing and reading the compact tries.